Moving `MeshContinuum::ExportCellsToExodus` into `MeshIO`
