+++ framework/mesh/io/vtk_io.cc
@@ -20,6 +20,8 @@
 #include <vtkXMLPUnstructuredGridReader.h>
 #include <vtkEnSightGoldBinaryReader.h>
 #include <vtkMultiBlockDataSet.h>
+#include <vtkExodusIIWriter.h>
+#include <vtkModelMetadata.h>
 #include <fstream>
 
 namespace opensn
@@ -1021,6 +1023,292 @@ MeshIO::ToOBJ(const std::shared_ptr<MeshContinuum>& grid, const char* file_name,
       log.Log() << "Exported Material Volume mesh to " << mat_file_name;
     } // for mat
   }   // if per material
+  opensn::mpi_comm.barrier();
+}
+
+void
+MeshIO::ToExodusII(const std::shared_ptr<MeshContinuum>& grid,
+                   const std::string& file_name,
+                   bool write_node_sets,
+                   bool write_side_sets)
+{
+  const std::string fname = "MeshIO::ToExodusII";
+  log.Log() << "Exporting mesh to ExodusII file with base " << file_name;
+
+  if (opensn::mpi_comm.size() != 1)
+    throw std::logic_error(fname + ": Currently this routine is only allowed in serial.");
+
+  // Check block consistency
+  std::map<int, CellType> block_id_map;
+  for (const auto& cell : grid->local_cells)
+  {
+    const int mat_id = cell.material_id_;
+    if (block_id_map.count(mat_id) == 0)
+      block_id_map[mat_id] = cell.SubType();
+    else
+    {
+      if (cell.SubType() != block_id_map.at(mat_id))
+        throw std::logic_error(fname + ": Material id " + std::to_string(mat_id) +
+                               " appearing for more than one cell type.");
+    }
+  }
+
+  // Create unstructured meshes for each material-type pair
+  vtkNew<vtkMultiBlockDataSet> grid_blocks;
+  int max_dimension = 0;
+  {
+    vtkNew<vtkUnstructuredGrid> ugrid;
+    vtkNew<vtkPoints> points;
+
+    points->SetDataType(VTK_DOUBLE);
+
+    vtkNew<vtkIdTypeArray> global_node_id_list;
+    global_node_id_list->SetName("GlobalNodeId");
+
+    vtkNew<vtkIdTypeArray> global_elem_id_list;
+    global_elem_id_list->SetName("GlobalElementId");
+
+    vtkNew<vtkIntArray> block_id_list;
+    block_id_list->SetName("BlockID");
+
+    // Load vertices
+    std::vector<uint64_t> vertex_map(grid->GetGlobalVertexCount(), 0);
+    const size_t num_verts = grid->GetGlobalVertexCount();
+    for (size_t v = 0; v < num_verts; ++v)
+    {
+      vertex_map[v] = v;
+      const auto& vertex = grid->vertices[v];
+      points->InsertNextPoint(vertex.x, vertex.y, vertex.z);
+
+      // Exodus node- and cell indices are 1-based therefore we add a 1 here.
+      global_node_id_list->InsertNextValue(static_cast<vtkIdType>(v + 1));
+    }
+
+    // Load cells
+    for (const auto& cell : grid->local_cells)
+    {
+      if (cell.SubType() == CellType::POLYGON or cell.SubType() == CellType::POLYHEDRON)
+        throw std::logic_error(fname + ": Cell-subtype \"" + CellTypeName(cell.SubType()) +
+                               "\" encountered that is not supported by ExodusII.");
+      UploadCellGeometryContinuous(cell, vertex_map, ugrid);
+      block_id_list->InsertNextValue(cell.material_id_);
+      max_dimension = std::max(max_dimension, MeshContinuum::GetCellDimension(cell));
+
+      // Exodus node- and cell indices are 1-based therefore we add a 1 here.
+      global_elem_id_list->InsertNextValue(static_cast<vtkIdType>(cell.global_id_ + 1));
+    } // for local cells
+
+    // Set arrays
+    ugrid->SetPoints(points);
+    ugrid->GetPointData()->AddArray(global_node_id_list);
+    ugrid->GetCellData()->AddArray(global_elem_id_list);
+    ugrid->GetCellData()->AddArray(block_id_list);
+
+    ugrid->GetPointData()->SetActiveGlobalIds("GlobalNodeId");
+    ugrid->GetCellData()->SetActiveGlobalIds("GlobalElementId");
+
+    // Set block
+    grid_blocks->SetBlock(0, ugrid);
+
+    log.Log() << "Writing grid block "
+              << " Number of cells: " << ugrid->GetNumberOfCells()
+              << " Number of points: " << ugrid->GetNumberOfPoints();
+  } // end of grid_blocks assignment
+
+  // Separate faces by boundary id
+  struct FaceInfo
+  {
+    CellFace const* face_ptr;
+    uint64_t source_cell_id;
+    int source_face_id;
+  };
+  std::map<uint64_t, std::vector<FaceInfo>> boundary_id_faces_map;
+  for (const auto& cell : grid->local_cells)
+  {
+    // Here we build a face mapping because OpenSn's face orientation for prisms (wedges) and
+    // hexahedrons differ from that of VTK. OpenSn's orientation for prisms and hexes actually
+    // matches that of Exodus but VTK assumes the incoming mesh to be conformant to VTK and
+    // therefore, internally performs a mapping. Fortunately, the only relevant cell-types, for
+    // which a special mapping is required, are the prisms and hexes.
+    const size_t num_faces = cell.faces_.size();
+    std::vector<int> face_mapping(num_faces, 0);
+    if (cell.SubType() == CellType::WEDGE)
+      face_mapping = {2, 3, 4, 0, 1};
+    else if (cell.SubType() == CellType::HEXAHEDRON)
+      face_mapping = {2, 1, 3, 0, 4, 5};
+    else
+    {
+      for (size_t f = 0; f < cell.faces_.size(); ++f)
+        face_mapping[f] = static_cast<int>(f);
+    }
+
+    // Here we store face information as a triplet, i.e., a face pointer, the id of the cell owning
+    // it, and the local face index (relative to the cell) of the face.
+    int f = 0;
+    for (const auto& face : cell.faces_)
+    {
+      if (not face.has_neighbor_)
+        boundary_id_faces_map[face.neighbor_id_].push_back(
+          {&face, cell.global_id_, face_mapping[f]});
+      ++f;
+    }
+  }
+
+  // Make NodeSets and/or SideSets
+  vtkNew<vtkMultiBlockDataSet> nodesets_blocks;
+  vtkNew<vtkMultiBlockDataSet> sidesets_blocks;
+  for (const auto& [bndry_id, face_list] : boundary_id_faces_map)
+  {
+    const std::string block_name = grid->GetBoundaryIDMap().at(bndry_id);
+    log.Log0Verbose1() << "bid: " + std::to_string(bndry_id) + " name=\"" + block_name + "\"";
+
+    // NodeSet
+    {
+      vtkNew<vtkUnstructuredGrid> ugrid;
+      vtkNew<vtkPoints> points;
+
+      points->SetDataType(VTK_DOUBLE);
+
+      vtkNew<vtkIdTypeArray> node_global_ids;
+      node_global_ids->SetName("GlobalNodeId");
+
+      // Build vertex set
+      std::set<uint64_t> vid_set;
+      for (const auto& face_info : face_list)
+        for (uint64_t vid : face_info.face_ptr->vertex_ids_)
+          vid_set.insert(vid);
+
+      // Build vertex map
+      std::vector<uint64_t> vertex_map(grid->GetGlobalVertexCount(), 0);
+      {
+        uint64_t mapped_id = 0;
+        for (uint64_t vid : vid_set)
+          vertex_map[vid] = mapped_id++;
+      }
+
+      // Load vertices
+      for (uint64_t vid : vid_set)
+      {
+        const auto& vertex = grid->vertices[vid];
+        points->InsertNextPoint(vertex.x, vertex.y, vertex.z);
+
+        // Exodus node- and cell indices are 1-based therefore we add a 1 here.
+        node_global_ids->InsertNextValue(static_cast<vtkIdType>(vid + 1));
+      }
+
+      // Load cells
+      for (uint64_t vid : vid_set)
+      {
+        std::vector<vtkIdType> cell_vids = {static_cast<vtkIdType>(vertex_map[vid])};
+        ugrid->InsertNextCell(VTK_VERTEX, static_cast<vtkIdType>(1), cell_vids.data());
+      }
+
+      ugrid->SetPoints(points);
+      ugrid->GetPointData()->AddArray(node_global_ids);
+
+      ugrid->GetPointData()->SetActiveGlobalIds("GlobalNodeId");
+
+      nodesets_blocks->SetBlock(bndry_id, ugrid);
+      nodesets_blocks->GetMetaData(bndry_id)->Set(vtkCompositeDataSet::NAME(), block_name);
+
+      log.Log() << "Writing node set block " << block_name
+                << " Number of cells: " << ugrid->GetNumberOfCells()
+                << " Number of points: " << ugrid->GetNumberOfPoints();
+    }
+
+    // SideSet
+    {
+      vtkNew<vtkUnstructuredGrid> ugrid;
+      vtkNew<vtkPoints> points;
+
+      points->SetDataType(VTK_DOUBLE);
+
+      vtkNew<vtkIdTypeArray> src_cell_global_ids;
+      vtkNew<vtkIntArray> src_cell_face_id;
+      src_cell_global_ids->SetName("SourceElementId");
+      src_cell_face_id->SetName("SourceElementSide");
+
+      // Build vertex set
+      std::set<uint64_t> vid_set;
+      for (const auto& face_info : face_list)
+        for (uint64_t vid : face_info.face_ptr->vertex_ids_)
+          vid_set.insert(vid);
+
+      // Build vertex map
+      std::vector<uint64_t> vertex_map(grid->GetGlobalVertexCount(), 0);
+      {
+        uint64_t mapped_id = 0;
+        for (uint64_t vid : vid_set)
+          vertex_map[vid] = mapped_id++;
+      }
+
+      // Load vertices
+      for (uint64_t vid : vid_set)
+      {
+        const auto& vertex = grid->vertices[vid];
+        points->InsertNextPoint(vertex.x, vertex.y, vertex.z);
+      }
+
+      // Load faces
+      for (const auto& face_info : face_list)
+      {
+        UploadFaceGeometry(*face_info.face_ptr, vertex_map, ugrid);
+        src_cell_global_ids->InsertNextValue(static_cast<vtkIdType>(face_info.source_cell_id));
+        src_cell_face_id->InsertNextValue(face_info.source_face_id);
+      }
+
+      ugrid->SetPoints(points);
+      ugrid->GetCellData()->AddArray(src_cell_global_ids);
+      ugrid->GetCellData()->AddArray(src_cell_face_id);
+
+      sidesets_blocks->SetBlock(bndry_id, ugrid);
+      sidesets_blocks->GetMetaData(bndry_id)->Set(vtkCompositeDataSet::NAME(), block_name);
+
+      log.Log() << "Writing side set block " << block_name
+                << " Number of cells: " << ugrid->GetNumberOfCells()
+                << " Number of points: " << ugrid->GetNumberOfPoints();
+    } // End of side-set
+  }
+
+  // Write the file
+  unsigned int next_block = 0;
+  vtkNew<vtkMultiBlockDataSet> main_block;
+  main_block->SetBlock(next_block++, grid_blocks);
+  if (write_node_sets)
+  {
+    log.Log0Verbose1() << "Exporting node set";
+    main_block->SetBlock(next_block, nodesets_blocks);
+    main_block->GetMetaData(next_block++)->Set(vtkCompositeDataSet::NAME(), "Node Sets");
+  }
+  if (write_side_sets)
+  {
+    log.Log0Verbose1() << "Exporting side set";
+    main_block->SetBlock(next_block, sidesets_blocks);
+    main_block->GetMetaData(next_block++)->Set(vtkCompositeDataSet::NAME(), "Side Sets");
+  }
+
+  vtkNew<vtkExodusIIWriter> writer;
+  writer->SetBlockIdArrayName("BlockID");
+
+  writer->SetFileName(file_name.c_str());
+  writer->SetStoreDoubles(1);
+
+  writer->SetInputData(main_block);
+
+  writer->WriteOutGlobalNodeIdArrayOff();
+  writer->WriteOutGlobalElementIdArrayOff();
+  writer->WriteOutBlockIdArrayOff();
+
+  writer->Write();
+
+  auto em = writer->GetModelMetadata();
+
+  log.Log() << "Num blocks   :  " << em->GetNumberOfBlocks();
+  log.Log() << "Num node sets:  " << em->GetNumberOfNodeSets();
+  log.Log() << "Num side sets:  " << em->GetNumberOfSideSets();
+  log.Log() << "Dimension    :  " << em->GetDimension();
+  log.Log() << "Done exporting mesh to ExodusII.";
+
   opensn::mpi_comm.barrier();
 }
